Remove pcrid-based broadcast filtering, and only normalise broadcast events if pcrid is present
